requirements.txt
pydub
azure-cognitiveservices-speech